requirements.txt
webex-bot [proxy]
atlassian-python-api
python-jenkins
